--- src/shared/types/index.ts
@@ -1,6 +1,0 @@
-import { SortHistory } from 'entities/sort_history';
-
-export type SortMapping = {
-  name: string;
-  value: (_array: number[]) => SortHistory;
-};
